refactor(hooks): group hooks by feature
Move app, favorite, nearby, route, routes, and shared hooks into feature folders under app/modules/hooks.

Update page, component, and test imports to use the new hook paths while preserving existing behavior.
