+++ app/components/nearby/NearbyStopRoutes.tsx
@@ -2,7 +2,7 @@ import { ScrollArea, Skeleton, Stack, Tabs, Text } from '@mantine/core'
 import { useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { SkeletonList } from '~/components/common/SkeletonList'
-import { useLocalizedTextCollator } from '~/modules/hooks/useLocalizedTextCollator'
+import { useLocalizedTextCollator } from '~/modules/hooks/shared/useLocalizedTextCollator'
 import { DirectionType } from '~/modules/enums/DirectionType'
 import type { StationRoute } from '~/modules/interfaces/StationRoute'
 import { getDirectionTranslationKey } from '~/modules/utils/i18n/getDirectionTranslationKey'
