+++ app/components/route/RouteStopList.tsx
@@ -4,7 +4,7 @@ import { useMemo, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
 import { getRouteRealtimeMessages } from '~/modules/consts/routeRealtimeMessages'
 import { RouteRealtimeInfoState } from '~/modules/enums/RouteRealtimeInfoState'
-import { useScrollSelectedItem } from '~/modules/hooks/useScrollSelectedItem'
+import { useScrollSelectedItem } from '~/modules/hooks/shared/useScrollSelectedItem'
 import type { FavoriteRouteStop } from '~/modules/interfaces/FavoriteRouteStop'
 import type { RouteRealtimeBusStatus } from '~/modules/interfaces/RouteRealtimeBusStatus'
 import { VehicleStateType } from '~/modules/enums/VehicleStateType'
