+++ app/modules/hooks/app/useGoogleAnalytics.ts
