+++ app/modules/hooks/nearby/useNearbyData.ts
@@ -19,7 +19,7 @@ import type { NearbyStopGroup } from '~/modules/interfaces/Nearby'
 import type { Stop } from '~/modules/interfaces/Stop'
 import type { StopOfRoute } from '~/modules/interfaces/StopOfRoute'
 import type { StationRoute } from '~/modules/interfaces/StationRoute'
-import { useLocalizedTextCollator } from '~/modules/hooks/useLocalizedTextCollator'
+import { useLocalizedTextCollator } from '~/modules/hooks/shared/useLocalizedTextCollator'
 import { selectLocale } from '~/modules/slices/localeSlice'
 import type { RootState } from '~/modules/store'
 import { getCityByCoords } from '~/modules/utils/geo/getCityByCoords'
