+++ app/modules/hooks/routes/useRouteSearchAnalytics.test.ts
