+++ app/modules/hooks/routes/useRouteSearchAnalytics.ts
