+++ app/pages/AppLayout.tsx
@@ -20,8 +20,8 @@ import {
   APP_FOOTER_HEIGHT,
   APP_HEADER_HEIGHT
 } from '~/modules/consts/layout'
-import { useWatchGeo } from '~/modules/hooks/useWatchGeo'
-import { useGoogleAnalytics } from '~/modules/hooks/useGoogleAnalytics'
+import { useWatchGeo } from '~/modules/hooks/app/useWatchGeo'
+import { useGoogleAnalytics } from '~/modules/hooks/app/useGoogleAnalytics'
 import { fetchCityGeoJSON } from '~/modules/slices/cityGeoSlice'
 import type { AppDispatch, RootState } from '~/modules/store'
