+++ app/pages/Favorite.tsx
@@ -6,8 +6,8 @@ import { BaseAlert } from '~/components/common/BaseAlert'
 import { FavoriteRouteStopCard } from '~/components/favorite/FavoriteRouteStopCard'
 import { APP_PAGE_PADDING } from '~/modules/consts/layout'
 import { getFavoriteMessages } from '~/modules/consts/pageMessages'
-import { useFavoriteRouteStops } from '~/modules/hooks/useFavoriteRouteStops'
-import { useLocalizedTextCollator } from '~/modules/hooks/useLocalizedTextCollator'
+import { useFavoriteRouteStops } from '~/modules/hooks/favorite/useFavoriteRouteStops'
+import { useLocalizedTextCollator } from '~/modules/hooks/shared/useLocalizedTextCollator'
 import { selectLocale } from '~/modules/slices/localeSlice'
 import { getLocalizedText } from '~/modules/utils/i18n/getLocalizedText'
