+++ app/pages/Nearby.tsx
@@ -4,9 +4,9 @@ import { useEffect, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
 import { MapSidebarLayout } from '~/components/common/MapSidebarLayout'
 import { NearbySidebarContent } from '~/components/nearby/NearbySidebarContent'
-import { useNearbyData } from '~/modules/hooks/useNearbyData'
-import { useScrollSelectedItem } from '~/modules/hooks/useScrollSelectedItem'
-import { useNearbySearchParams } from '~/modules/hooks/useNearbySearchParams'
+import { useNearbyData } from '~/modules/hooks/nearby/useNearbyData'
+import { useScrollSelectedItem } from '~/modules/hooks/shared/useScrollSelectedItem'
+import { useNearbySearchParams } from '~/modules/hooks/nearby/useNearbySearchParams'
 import { NearbyStopDetail } from '~/components/nearby/NearbyStopDetail'
 import { NearbyStopMap } from '~/components/nearby/NearbyStopMap'
 import { RiMenuFill } from '@remixicon/react'
