+++ app/pages/Route.tsx
@@ -8,10 +8,10 @@ import { MapSidebarLayout } from '~/components/common/MapSidebarLayout'
 import { useLocation, useNavigate, useParams } from 'react-router'
 import { RouteMap } from '~/components/route/RouteMap'
 import { RouteStopList } from '~/components/route/RouteStopList'
-import { useFavoriteRouteStops } from '~/modules/hooks/useFavoriteRouteStops'
-import { useRouteBaseData } from '~/modules/hooks/useRouteBaseData'
-import { useRouteRealtimeData } from '~/modules/hooks/useRouteRealtimeData'
-import { useRouteSelectionState } from '~/modules/hooks/useRouteSelectionState'
+import { useFavoriteRouteStops } from '~/modules/hooks/favorite/useFavoriteRouteStops'
+import { useRouteBaseData } from '~/modules/hooks/route/useRouteBaseData'
+import { useRouteRealtimeData } from '~/modules/hooks/route/useRouteRealtimeData'
+import { useRouteSelectionState } from '~/modules/hooks/route/useRouteSelectionState'
 import { RiArrowLeftSLine, RiMenuFill } from '@remixicon/react'
 import { AppBadge } from '~/components/common/AppBadge'
 import { selectLocale } from '~/modules/slices/localeSlice'
